requirements.txt
fastapi
uvicorn
mysql-connector-python
transformers
torch
Jinja2
sentencepiece
python-dotenv